Split up `TableHeader.tsx` into 3 files
